refactor: update rich text section
