+++ app/sections/RichText/RichText.tsx
@@ -9,24 +9,23 @@ interface RichTextProps extends HydrogenComponentProps {
   contentAlignment: string;
   sectionHeight: string;
   backgroundColor: string;
-  textColor: string;
   topPadding: string;
   bottomPadding: string;
 }
 
 let RichText = forwardRef<HTMLElement, RichTextProps>((props, ref) => {
-  let { contentAlignment, sectionHeight, backgroundColor, textColor, topPadding, bottomPadding, children, ...rest } = props;
+  let { contentAlignment, sectionHeight, backgroundColor, topPadding, bottomPadding, children, ...rest } = props;
   let sectionStyle: CSSProperties = {
     alignItems: `${contentAlignment}`,
     '--section-height': `${sectionHeight}px`,
     backgroundColor: `${backgroundColor}`,
-    '--text-color': `${textColor}`,
     paddingTop: `${topPadding}px`,
     paddingBottom: `${bottomPadding}px`,
+    '--max-width-content': '600px'
   } as CSSProperties;
   return (
     <section ref={ref} {...rest} className='py-16 px-10 h-[var(--section-height)] flex flex-col justify-center' style={sectionStyle}>
-      <div className='text-center w-full flex flex-col gap-5 sm:w-1/2'>
+      <div className='text-center w-full flex flex-col gap-5 sm:w-[var(--max-width-content)]'>
         {children}
       </div>
     </section>
@@ -48,12 +47,6 @@ export let schema: HydrogenComponentSchema = {
           name: 'backgroundColor',
           label: 'Background color',
           defaultValue: '#F7F7F7',
-        },
-        {
-          type: 'color',
-          name: 'textColor',
-          label: 'Text color',
-          defaultValue: '#000000',
         },
         {
           type: 'toggle-group',
@@ -107,14 +100,16 @@ export let schema: HydrogenComponentSchema = {
       ],
     },
   ],
-  childTypes: ['rich-text-heading--item', 'rich-text-description--item', 'rich-text-button--item'],
+  childTypes: ['heading', 'description', 'rich-text-button--item'],
   presets: {
     children: [
       {
-        type: 'rich-text-heading--item',
+        type: 'heading',
+        content: 'Rich text',
       },
       {
-        type: 'rich-text-description--item',
+        type: 'description',
+        content: 'Pair large text with an image to tell a story, explain a detail about your product, or describe a new promotion.',
       },
       {
         type: 'rich-text-button--item',
